refactor(web): app に閉じる hooks を app/<feature>/hooks/ に colocate する
dashboard/chat 内でのみ使用される hooks を ~/hooks/ から各 app 配下に移動。
複数 app から使う useAdminUser / useCopyToClipboard のみ ~/hooks/ に残す。
import は同一階層は相対、それ以外は ~ alias に統一。
