+++ web/src/app/chat/hooks/useAnonymousSession.test.ts
