+++ web/src/app/chat/hooks/useAnonymousSession.ts
