+++ web/src/app/dashboard/components/InvitationsPanel.tsx
@@ -5,9 +5,9 @@ import {
   useCreateInvitation,
   useDeleteInvitation,
   useInvitations,
-} from "~/hooks/dashboard/useInvitations";
+} from "~/app/dashboard/hooks/useInvitations";
+import { ROLE_LABELS, useRole } from "~/app/dashboard/hooks/useRole";
 import { useCopyToClipboard } from "~/hooks/useCopyToClipboard";
-import { ROLE_LABELS, useRole } from "~/hooks/useRole";
 import type { AdminUser } from "~/lib/api/auth";
 import { formatDateTime } from "~/lib/format";
 import { buildInvitationUrl, isExpired } from "./invitation/helpers";
