+++ web/src/app/dashboard/components/PersonaPanel.tsx
@@ -1,9 +1,9 @@
+import { useInfiniteScroll } from "~/app/dashboard/hooks/useInfiniteScroll";
 import {
   useDeletePersonas,
   useExtractPersonas,
   usePersonas,
-} from "~/hooks/dashboard/usePersonas";
-import { useInfiniteScroll } from "~/hooks/useInfiniteScroll";
+} from "~/app/dashboard/hooks/usePersonas";
 import { confirmDialog } from "~/lib/dialog";
 import { formatDateTime } from "~/lib/format";
 import { getSentimentStyle } from "./persona/helpers";
