+++ web/src/app/dashboard/hooks/useEmergencies.test.ts
