+++ web/src/app/dashboard/hooks/useEmergencies.ts
